Feature: Clean no more used gems  and code (#658)
* remove cube ruby

* remove not used gems and add documentation to the Gemfile

* remove miniprofiler

* remove bpdi resolver and redirect_to_new_api

* remove ajax proxy and no more used helpers

* remove no more used helpers
