+++ app/controllers/annotator_controller.rb
@@ -73,7 +73,7 @@ class AnnotatorController < ApplicationController
         @results_table_header.push(t('annotator.score'))
       end
       annotations = LinkedData::Client::HTTP.get(uri.dup.to_s, api_params)
-      @ontologies = get_simplified_ontologies_hash
+      @ontologies = LinkedData::Client::Models::Ontology.all({:include_views => true}).map{ |o| [o.id.to_s, o]}.to_h
       @semantic_types = get_semantic_types 
       @results = []
       annotations.each do |annotation|
@@ -161,14 +161,14 @@ class AnnotatorController < ApplicationController
   end
 
   def annotation_class_info(cls)
-    return {
+    {
       text: cls.prefLabel,
       link: url_to_endpoint(cls.links["self"])
     }
   end
   def annotation_ontology_info(ontology_url)
-    return {
-      text: @ontologies[ontology_url][:name],
+    {
+      text: @ontologies[ontology_url].name,
       link: url_to_endpoint(ontology_url)
     }
   end
