+++ app/controllers/application_controller.rb
@@ -53,15 +53,6 @@ class ApplicationController < ActionController::Base
   # Pull configuration parameters for REST connection.
   REST_URI = $REST_URL
   API_KEY = $API_KEY
-  PROXY_URI = $PROXY_URL
-  REST_URI_BATCH = REST_URI + '/batch'
-
-  # Rails.cache expiration
-  EXPIRY_RI_STATS = 60 * 60 * 24       # 24:00 hours
-  EXPIRY_RI_ONTOLOGIES = 60 * 60 * 24  # 24:00 hours
-  EXPIRY_SEMANTIC_TYPES = 60 * 60 * 24 # 24:00 hours
-  EXPIRY_RECENT_MAPPINGS = 60 * 60     #  1:00 hours
-  EXPIRY_ONTOLOGY_SIMPLIFIED = 60 * 1  #  0:01 minute
 
 
   $DATA_CATALOG_VALUES = {"fairsharing.org/" => "FAIRsharing",
@@ -90,13 +81,12 @@ class ApplicationController < ActionController::Base
   # See ActionController::RequestForgeryProtection for details
   protect_from_forgery
 
-  before_action :set_global_thread_values, :domain_ontology_set, :authorize_miniprofiler, :clean_empty_strings_from_params_arrays, :init_trial_license
+  before_action :set_global_thread_values, :domain_ontology_set, :clean_empty_strings_from_params_arrays, :init_trial_license
 
 
   def invalidate_cache?
     params[:invalidate_cache] && params[:invalidate_cache].to_s.eql?('true')
   end
-
   def show_image_modal
     url = params[:url]
     render turbo_stream: helpers.prepend('application_modal_content') { helpers.image_tag(url, style:'width: 100%') }
@@ -139,8 +129,6 @@ class ApplicationController < ActionController::Base
     Thread.current[:slice] = @subdomain_filter
   end
 
-
-
   def ontology_not_found(ontology_acronym)
     not_found(t('application.ontology_not_found',acronym: ontology_acronym))
   end
@@ -156,14 +144,6 @@ class ApplicationController < ActionController::Base
     end
 
     raise ActiveRecord::RecordNotFound.new(message || t('application.not_found_message'))
-  end
-
-  NOTIFICATION_TYPES = { :notes => "CREATE_NOTE_NOTIFICATION", :all => "ALL_NOTIFICATION" }
-
-  def to_param(name) # Paramaterizes URLs without encoding
-    unless name.nil?
-      name.to_s.gsub(' ',"_")
-    end
   end
 
   def bp_config_json
@@ -191,36 +171,8 @@ class ApplicationController < ActionController::Base
     config.to_json
   end
 
-
   def rest_url
     helpers.rest_url
-  end
-  
-  def check_http_file(url)
-    session = Net::HTTP.new(url.host, url.port)
-    session.use_ssl = true if url.port == 443
-    session.start do |http|
-      response_valid = http.head(url.request_uri).code.to_i < 400
-      return response_valid
-    end
-  end
-
-  def check_ftp_file(uri)
-    ftp = Net::FTP.new(uri.host, uri.user, uri.password)
-    ftp.login
-    begin
-      file_exists = ftp.size(uri.path) > 0
-    rescue
-      # Check using another method
-      path = uri.path.split("/")
-      filename = path.pop
-      path = path.join("/")
-      ftp.chdir(path)
-      files = ftp.dir
-      # Dumb check, just see if the filename is somewhere in the list
-      files.each { |file| return true if file.include?(filename) }
-    end
-    file_exists
   end
 
   def parse_response_body(response)
@@ -285,62 +237,12 @@ class ApplicationController < ActionController::Base
     redirect_to "/"
   end
 
-
-  def redirect_new_api(class_view = false)
-    # Hack to make ontologyid and conceptid work in addition to id and ontology params
-    params[:ontology] = params[:ontology].nil? ? params[:ontologyid] : params[:ontology]
-    # Error checking
-    if params[:ontology].nil? || params[:id] && params[:ontology].nil?
-      @error = t('application.provide_ontology_or_concept')
-      return
-    end
-    acronym = BpidResolver.id_to_acronym(params[:ontology])
-    ontology_not_found(params[:ontology]) unless acronym
-    if class_view
-      @ontology = LinkedData::Client::Models::Ontology.find_by_acronym(acronym).first
-      concept = get_class(params).first.to_s
-      redirect_to "/ontologies/#{acronym}?p=classes#{params_string_for_redirect(params, prefix: "&")}", :status => :moved_permanently
-    else
-      redirect_to "/ontologies/#{acronym}#{params_string_for_redirect(params)}", :status => :moved_permanently
-    end
-  end
-
-  def params_cleanup_new_api
-    params = @_params
-    if params[:ontology] && params[:ontology].to_i > 0
-      params[:ontology] = BpidResolver.id_to_acronym(params[:ontology])
-    end
-
-    params
-  end
-
-  def params_string_for_redirect(params, options = {})
-    prefix = options[:prefix] || "?"
-    stop_words = options[:stop_words] || ["ontology", "controller", "action", "id", "acronym"]
-    params_array = []
-    params.each do |key,value|
-      next if stop_words.include?(key.to_s) || value.nil? || value.empty?
-      params_array << "#{key}=#{CGI.escape(value)}"
-    end
-    params_array.empty? ? "" : "#{prefix}#{params_array.join('&')}"
-  end
-
-  # rack-mini-profiler authorization
-  def authorize_miniprofiler
-    if params[:enable_profiler] && params[:enable_profiler].eql?("true") && session[:user] && session[:user].admin?
-      Rack::MiniProfiler.authorize_request
-    else
-      Rack::MiniProfiler.deauthorize_request
-    end
-  end
-
   # Verifies if user is logged in
   def authorize_and_redirect
     unless session[:user]
       redirect_to_home
     end
   end
-
 
   def authorize_admin
     admin = session[:user] && session[:user].admin?
@@ -355,7 +257,6 @@ class ApplicationController < ActionController::Base
     restrict_downloads = $NOT_DOWNLOADABLE
     restrict_downloads.include? acronym
   end
-
 
   def check_delete_mapping_permission(mappings)
     # ensure mappings is an Array of mappings (some calls may provide only a single mapping instance)
@@ -465,7 +366,6 @@ class ApplicationController < ActionController::Base
     @concept
   end
 
-
   def get_ontology_submission_ready(ontology)
     # Get the latest 'ready' submission
     submission = ontology.explore.latest_submission({:include_status => 'ready'})
@@ -474,190 +374,12 @@ class ApplicationController < ActionController::Base
     return submission
   end
 
-  def get_simplified_ontologies_hash()
-    # Note the simplify_ontology_model will cache individual ontology data.
-    simple_ontologies = {}
-    begin
-      ontology_models = LinkedData::Client::Models::Ontology.all({:include_views => true})
-      ontology_models.each {|o| simple_ontologies[o.id] = simplify_ontology_model(o) }
-    rescue Exception => e
-      LOG.add :error, e.message
-      return nil
-    end
-    return simple_ontologies
-  end
-
-
-
-  def simplify_classes(classes)
-    # Simplify the classes batch service data for the UI
-    # It takes a list of class objects (hashes or models) and the
-    # data structure returned is a hash of class hashes, which will
-    # contain details for the ontology they belong to.  For example:
-    #{
-    # "http://ncicb.nci.nih.gov/xml/owl/EVS/Thesaurus.owl#C12439" => {
-    #    :id => "http://ncicb.nci.nih.gov/xml/owl/EVS/Thesaurus.owl#C12439",
-    #    :ui => "http://ncbo-stg-app-12.stanford.edu/ontologies/NCIT?p=classes&conceptid=http%3A%2F%2Fncicb.nci.nih.gov%2Fxml%2Fowl%2FEVS%2FThesaurus.owl%23C12439",
-    #    :uri => "http://stagedata.bioontology.org/ontologies/NCIT/classes/http%3A%2F%2Fncicb.nci.nih.gov%2Fxml%2Fowl%2FEVS%2FThesaurus.owl%23C12439",
-    #    :prefLabel => "Brain",
-    #    :ontology => {
-    #      :id => "http://stagedata.bioontology.org/ontologies/NCIT",
-    #      :uri => "http://stagedata.bioontology.org/ontologies/NCIT",
-    #      :acronym => "NCIT",
-    #      :name => "National Cancer Institute Thesaurus",
-    #      :ui => "http://ncbo-stg-app-12.stanford.edu/ontologies/NCIT"
-    #    },
-    #  },
-    #}
-    @ontologies_hash ||= get_simplified_ontologies_hash
-    classes_hash = {}
-    classes.each do |cls|
-      c = simplify_class_model(cls)
-      c[:ontology] = @ontologies_hash[ c[:ontology] ]
-      classes_hash[c[:id]] = c
-    end
-    return classes_hash
-  end
-
-  def simplify_class_model(cls_model)
-    # Simplify the class required required by the UI.
-    # No modification of the class ontology here, see simplify_classes.
-    # Default simple class model
-    cls = { :id => nil, :ontology => nil, :prefLabel => nil, :uri => nil, :ui => nil, :obsolete => false }
-    begin
-      if cls_model.instance_of? Hash
-        cls = {
-            :id => cls_model['@id'],
-            :ui =>  cls_model['links']['ui'],
-            :uri => cls_model['links']['self'],  # different from id
-            :ontology => cls_model['links']['ontology']
-        }
-        # Try to carry through a prefLabel and the obsolete attribute, if they exist.
-        cls[:prefLabel] = cls_model['prefLabel']
-        cls[:obsolete] = cls_model['obsolete'] || false
-      else
-        # try to work with a struct object or a LinkedData::Client::Models::Class
-        # if not a struct, then: cls_model.instance_of? LinkedData::Client::Models::Class
-        cls = {
-            :id => cls_model.id,
-            :ui =>  cls_model.links['ui'],
-            :uri => cls_model.links['self'],  # different from id
-            :ontology => cls_model.links['ontology'],
-        }
-        # Try to carry through a prefLabel and the obsolete attribute, if they exist.
-        cls[:prefLabel] = cls_model.prefLabel if cls_model.respond_to?('prefLabel')
-        cls[:obsolete] = cls_model.respond_to?('obsolete') && cls_model.obsolete || false
-      end
-    rescue Exception => e
-      LOG.add :error, e.message
-      LOG.add :error, t('application.error_simplify_class', cls: cls)
-    end
-    return cls
-  end
-
-  def simplify_ontology_model(ont_model)
-    id = nil
-    if ont_model.instance_of? Hash
-      id = ont_model['@id']
-    elsif ont_model.instance_of? LinkedData::Client::Models::Ontology
-      id = ont_model.id
-    end
-    ont = Rails.cache.read(id)
-    return ont unless ont.nil?
-    # No cache or it has expired
-    LOG.add :debug, t('application.no_cache', id: id)
-    ont = {}
-    ont[:id] = id
-    ont[:uri] = id
-    if ont_model.instance_of? Hash
-      ont[:acronym] = ont_model['acronym']
-      ont[:name] = ont_model['name']
-      ont[:ui] = ont_model['links']['ui']
-    else
-      # try to work with a struct object or a LinkedData::Client::Models::Ontology
-      # if not a struct, then: ont_model.instance_of? LinkedData::Client::Models::Ontology
-      ont[:acronym] = ont_model.acronym
-      ont[:name] = ont_model.name
-      ont[:ui] = ont_model.links['ui']
-    end
-    # Only cache a complete representation of a simplified ontology
-    if ont[:id].nil? || ont[:uri].nil? || ont[:acronym].nil? || ont[:name].nil? || ont[:ui].nil?
-      raise t('application.incomplete_simple_ontology', id: id, ont: ont)
-    else
-      Rails.cache.write(ont[:id], ont, expires_in: EXPIRY_ONTOLOGY_SIMPLIFIED)
-    end
-    return ont
-  end
-
   def get_apikey()
     apikey = API_KEY
     if session[:user]
       apikey = session[:user].apikey
     end
     return apikey
-  end
-
-  def parse_json(uri)
-    uri = URI.parse(uri)
-    begin
-      response = open(uri, "Authorization" => "apikey token=#{get_apikey}").read
-    rescue Exception => error
-      @retries ||= 0
-      if @retries < 1  # retry once only
-        @retries += 1
-        retry
-      else
-        raise error
-      end
-    end
-    JSON.parse(response)
-  end
-
-
-  def get_batch_results(params)
-    begin
-      response = RestClient.post REST_URI_BATCH, params.to_json, :content_type => :json, :accept => :json, :authorization => "apikey token=#{get_apikey}"
-    rescue Exception => error
-      @retries ||= 0
-      if @retries < 1  # retry once only
-        @retries += 1
-        retry
-      else
-        LOG.add :error, "\nERROR: batch POST, uri: #{REST_URI_BATCH}"
-        LOG.add :error, "\nERROR: batch POST, params: #{params.to_json}"
-        LOG.add :error, "\nERROR: batch POST, error response: #{error.response}"
-        raise error
-      end
-    end
-    response
-  end
-
-  # Get the latest manual mappings
-  # All mapping classes are bidirectional.
-  # Each class in the list maps to all other classes in the list.
-  def get_recent_mappings
-    recent_mappings = {
-        :mappings => [],
-        :classes => {}
-    }
-    begin
-      recent_url = "#{REST_URI}/mappings/recent/"
-      cached_mappings_key = recent_url
-      cached_mappings = Rails.cache.read(cached_mappings_key)
-      return cached_mappings unless (cached_mappings.nil? || cached_mappings.empty?)
-      # No cache or it has expired
-      class_details = {}
-      mappings = LinkedData::Client::HTTP.get(recent_url, {size: 20, display: "prefLabel"})
-      recent_mappings[:mappings] = mappings
-      unless mappings.nil? || mappings.empty?
-        # Only cache a successful retrieval
-        Rails.cache.write(cached_mappings_key, recent_mappings, expires_in: EXPIRY_RECENT_MAPPINGS)
-      end
-    rescue Exception => e
-      LOG.add :error, e.message
-      # leave recent mappings empty.
-    end
-    return recent_mappings
   end
 
   def total_mapping_count
