+++ app/controllers/ontologies_controller.rb
@@ -232,17 +232,10 @@ class OntologiesController < ApplicationController
       else
         params[:conceptid] = params.delete(:purl_conceptid)
       end
-      redirect_to "/ontologies/#{params[:acronym]}?p=classes#{params_string_for_redirect(params, prefix: "&")}", status: :moved_permanently
+      redirect_to "/ontologies/#{params[:acronym]}?p=classes&conceptid=#{params[:conceptid]}", status: :moved_permanently
       return
     end
 
-    if params[:ontology].to_i > 0
-      acronym = BPIDResolver.id_to_acronym(params[:ontology])
-      if acronym
-        redirect_new_api
-        return
-      end
-    end
 
     # Note: find_by_acronym includes ontology views
     @ontology = LinkedData::Client::Models::Ontology.find_by_acronym(params[:ontology]).first
@@ -279,9 +272,6 @@ class OntologiesController < ApplicationController
 
     # This action is now a router using the 'p' parameter as the page to show
     case params[:p]
-    when 'terms'
-      params[:p] = 'classes'
-      redirect_to "/ontologies/#{params[:ontology]}#{params_string_for_redirect(params)}", status: :moved_permanently
     when 'classes'
       self.classes # rescue self.summary
     when 'mappings'
@@ -386,13 +376,6 @@ class OntologiesController < ApplicationController
     }
   end
 
-  def virtual
-    redirect_new_api
-  end
-
-  def visualize
-    redirect_new_api(true)
-  end
 
   def widgets
     if request.xhr?
