+++ app/views/mappings/_form.html.haml
@@ -3,7 +3,7 @@
 
 %div
   %span#create_mapping_spinner.card.p-3.text-center
-    = loading_spinner
+    = loader_component
 
   = render_alerts_container
