--- config/initializers/miniprofiler.rb
@@ -1,2 +1,0 @@
-Rack::MiniProfiler.config.authorization_mode = :allow_authorized
-Rack::MiniProfiler.config.position = 'right'
