--- lib/resolver/acronym_from_virtual.rb
@@ -1,535 +1,0 @@
-BpidResolver::ACRONYM_FROM_VIRTUAL = {
-  "old:acronym_from_virtual:3013" => "DIAGONT",
-  "old:acronym_from_virtual:3245" => "SuicidO",
-  "old:acronym_from_virtual:1247" => "GEOSPECIES",
-  "old:acronym_from_virtual:1054" => "AMINO-ACID",
-  "old:acronym_from_virtual:2012" => "goslim_pir",
-  "old:acronym_from_virtual:3015" => "PMR",
-  "old:acronym_from_virtual:3212" => "GLYCOPROT",
-  "old:acronym_from_virtual:3271" => "BSAO",
-  "old:acronym_from_virtual:2099" => "GWAS_EFO_SKOS",
-  "old:acronym_from_virtual:1062" => "PR",
-  "old:acronym_from_virtual:3004" => "NATPRO",
-  "old:acronym_from_virtual:1584" => "MMO",
-  "old:acronym_from_virtual:1046" => "SBO",
-  "old:acronym_from_virtual:1545" => "CO-WHEAT",
-  "old:acronym_from_virtual:1016" => "FB-DV",
-  "old:acronym_from_virtual:1059" => "CPRO",
-  "old:acronym_from_virtual:3238" => "PDO",
-  "old:acronym_from_virtual:1576" => "FDA-MEDDEVICE",
-  "old:acronym_from_virtual:1114" => "BILA",
-  "old:acronym_from_virtual:1055" => "GALEN",
-  "old:acronym_from_virtual:1158" => "CBO",
-  "old:acronym_from_virtual:3151" => "GEXO",
-  "old:acronym_from_virtual:2109" => "ooevv-vaccine",
-  "old:acronym_from_virtual:1488" => "JERM",
-  "old:acronym_from_virtual:3234" => "BOF",
-  "old:acronym_from_virtual:3047" => "DEMOGRAPH",
-  "old:acronym_from_virtual:1613" => "BDO",
-  "old:acronym_from_virtual:1142" => "EP",
-  "old:acronym_from_virtual:3127" => "IMMDIS",
-  "old:acronym_from_virtual:1461" => "TMO",
-  "old:acronym_from_virtual:3146" => "CMS",
-  "old:acronym_from_virtual:1125" => "HP",
-  "old:acronym_from_virtual:2033" => "iceci-instrument_object_substance",
-  "old:acronym_from_virtual:1033" => "NMR",
-  "old:acronym_from_virtual:1666" => "MFOEM",
-  "old:acronym_from_virtual:1615" => "CHEMBIO",
-  "old:acronym_from_virtual:3184" => "ZIP3",
-  "old:acronym_from_virtual:3150" => "BRIDG",
-  "old:acronym_from_virtual:2054" => "NCBI_NMOsp_1_0_0",
-  "old:acronym_from_virtual:1370" => "ATO",
-  "old:acronym_from_virtual:3268" => "ERNO",
-  "old:acronym_from_virtual:3261" => "STNFRDRXDEMO",
-  "old:acronym_from_virtual:2032" => "iceci-details_for_place_of_occurrence",
-  "old:acronym_from_virtual:1402" => "NIFCELL",
-  "old:acronym_from_virtual:1427" => "RCD",
-  "old:acronym_from_virtual:1565" => "OMRSE",
-  "old:acronym_from_virtual:1517" => "EHDAA2",
-  "old:acronym_from_virtual:1347" => "MEDLINEPLUS",
-  "old:acronym_from_virtual:3243" => "NHSQI2009",
-  "old:acronym_from_virtual:1640" => "PEDTERM",
-  "old:acronym_from_virtual:1354" => "WHO-ART",
-  "old:acronym_from_virtual:1095" => "XAO",
-  "old:acronym_from_virtual:1661" => "UCSFXPLANT",
-  "old:acronym_from_virtual:3174" => "IDODEN",
-  "old:acronym_from_virtual:1146" => "ECG",
-  "old:acronym_from_virtual:1065" => "TADS",
-  "old:acronym_from_virtual:3028" => "SOY",
-  "old:acronym_from_virtual:1100" => "OGI",
-  "old:acronym_from_virtual:2086" => "SMD_NPLSM_BRST",
-  "old:acronym_from_virtual:1136" => "EFO",
-  "old:acronym_from_virtual:1050" => "ZEA",
-  "old:acronym_from_virtual:3124" => "VSO",
-  "old:acronym_from_virtual:2085" => "SMD_NPLSM",
-  "old:acronym_from_virtual:1047" => "GRO-CPD",
-  "old:acronym_from_virtual:1540" => "DDI",
-  "old:acronym_from_virtual:2007" => "goslim_generic_1_770",
-  "old:acronym_from_virtual:3155" => "ONTOKBCF",
-  "old:acronym_from_virtual:2045" => "ICF-d5",
-  "old:acronym_from_virtual:1350" => "LOINC",
-  "old:acronym_from_virtual:1099" => "ATMO",
-  "old:acronym_from_virtual:3231" => "GLYCANONT",
-  "old:acronym_from_virtual:3131" => "PROVO",
-  "old:acronym_from_virtual:1123" => "OBI",
-  "old:acronym_from_virtual:2035" => "iceci-mechanism",
-  "old:acronym_from_virtual:1420" => "MSHSPA_1",
-  "old:acronym_from_virtual:1304" => "BCGO",
-  "old:acronym_from_virtual:1042" => "OBOREL",
-  "old:acronym_from_virtual:1049" => "WB-LS",
-  "old:acronym_from_virtual:1328" => "HOM",
-  "old:acronym_from_virtual:1575" => "EXO",
-  "old:acronym_from_virtual:3077" => "ONTODM-KDD",
-  "old:acronym_from_virtual:2108" => "ooevv-tractTrace",
-  "old:acronym_from_virtual:3191" => "MPO",
-  "old:acronym_from_virtual:1349" => "PDQ",
-  "old:acronym_from_virtual:3239" => "CSSO",
-  "old:acronym_from_virtual:2038" => "iceci-details_for_activity",
-  "old:acronym_from_virtual:1522" => "BP",
-  "old:acronym_from_virtual:1587" => "PO",
-  "old:acronym_from_virtual:1116" => "BHO",
-  "old:acronym_from_virtual:3038" => "HOM-TEST",
-  "old:acronym_from_virtual:1001" => "GRO-CPGA",
-  "old:acronym_from_virtual:2025" => "NIF-RTH",
-  "old:acronym_from_virtual:1537" => "IDOBRU",
-  "old:acronym_from_virtual:1009" => "DOID",
-  "old:acronym_from_virtual:3178" => "CANONT",
-  "old:acronym_from_virtual:2036" => "iceci-place_of_occurrence",
-  "old:acronym_from_virtual:1112" => "UO",
-  "old:acronym_from_virtual:3153" => "NTDO",
-  "old:acronym_from_virtual:1010" => "EMAP",
-  "old:acronym_from_virtual:1541" => "MCCL",
-  "old:acronym_from_virtual:1067" => "WB-PHENOTYPE",
-  "old:acronym_from_virtual:1560" => "COGPO",
-  "old:acronym_from_virtual:1697" => "UCSFORTHO",
-  "old:acronym_from_virtual:1260" => "RID",
-  "old:acronym_from_virtual:1091" => "SPD",
-  "old:acronym_from_virtual:2103" => "ICPCFRE",
-  "old:acronym_from_virtual:1069" => "ENVO",
-  "old:acronym_from_virtual:1686" => "NEOMARK4",
-  "old:acronym_from_virtual:1527" => "VANDF",
-  "old:acronym_from_virtual:1057" => "RADLEX",
-  "old:acronym_from_virtual:3258" => "OBR-Scolio",
-  "old:acronym_from_virtual:1500" => "RNAO",
-  "old:acronym_from_virtual:3042" => "TEO",
-  "old:acronym_from_virtual:2114" => "citation",
-  "old:acronym_from_virtual:2003" => "BiositemapIM",
-  "old:acronym_from_virtual:2029" => "iceci-countermeasures",
-  "old:acronym_from_virtual:3264" => "WH",
-  "old:acronym_from_virtual:1144" => "DC-CL",
-  "old:acronym_from_virtual:1418" => "TOK",
-  "old:acronym_from_virtual:1497" => "PMA",
-  "old:acronym_from_virtual:2093" => "OBI_IEDB_view",
-  "old:acronym_from_virtual:3252" => "SSE",
-  "old:acronym_from_virtual:1183" => "LIPRO",
-  "old:acronym_from_virtual:1423" => "RXNORM",
-  "old:acronym_from_virtual:1026" => "MAO",
-  "old:acronym_from_virtual:1555" => "VSAO",
-  "old:acronym_from_virtual:3031" => "QUDT",
-  "old:acronym_from_virtual:3223" => "GCO",
-  "old:acronym_from_virtual:1172" => "VO",
-  "old:acronym_from_virtual:1068" => "SAO",
-  "old:acronym_from_virtual:1410" => "KISAO",
-  "old:acronym_from_virtual:1520" => "LSM",
-  "old:acronym_from_virtual:2050" => "ICF-d810-d839",
-  "old:acronym_from_virtual:3021" => "ATC",
-  "old:acronym_from_virtual:1063" => "CARO",
-  "old:acronym_from_virtual:1530" => "ADW",
-  "old:acronym_from_virtual:2120" => "provenance",
-  "old:acronym_from_virtual:3017" => "GCC",
-  "old:acronym_from_virtual:2056" => "NCBI_NMOsp_1_2_2",
-  "old:acronym_from_virtual:3206" => "HCPCS-HIMC",
-  "old:acronym_from_virtual:1104" => "BRO",
-  "old:acronym_from_virtual:1092" => "IDO",
-  "old:acronym_from_virtual:1413" => "SWO",
-  "old:acronym_from_virtual:3032" => "EPICMEDS",
-  "old:acronym_from_virtual:1456" => "MSHFRE",
-  "old:acronym_from_virtual:1077" => "MIRO",
-  "old:acronym_from_virtual:1051" => "ZFA",
-  "old:acronym_from_virtual:2055" => "NCBI_NMOsp_1_2_1",
-  "old:acronym_from_virtual:3090" => "ONTOMA",
-  "old:acronym_from_virtual:3104" => "USSOC",
-  "old:acronym_from_virtual:1573" => "TM-MER",
-  "old:acronym_from_virtual:3092" => "CLINIC",
-  "old:acronym_from_virtual:1616" => "PHYLONT",
-  "old:acronym_from_virtual:2023" => "NCBITaxonOWL",
-  "old:acronym_from_virtual:1656" => "PHENOMEBLAST",
-  "old:acronym_from_virtual:3219" => "MSTDE",
-  "old:acronym_from_virtual:1025" => "MP",
-  "old:acronym_from_virtual:1670" => "ICDO3",
-  "old:acronym_from_virtual:1089" => "BIRNLEX",
-  "old:acronym_from_virtual:3215" => "MIXS",
-  "old:acronym_from_virtual:1532" => "SIO",
-  "old:acronym_from_virtual:2105" => "SNMD_BDY",
-  "old:acronym_from_virtual:3022" => "BIOMODELS",
-  "old:acronym_from_virtual:2053" => "ICF-d840-d859%2Fd860-d879",
-  "old:acronym_from_virtual:3262" => "BCTEO",
-  "old:acronym_from_virtual:1168" => "Amino_Acid-zh_CN",
-  "old:acronym_from_virtual:1550" => "PHARE",
-  "old:acronym_from_virtual:1398" => "LDA",
-  "old:acronym_from_virtual:1428" => "NIC",
-  "old:acronym_from_virtual:1567" => "PVONTO",
-  "old:acronym_from_virtual:3211" => "TEST-PROD",
-  "old:acronym_from_virtual:2124" => "ICD9CM-PROC",
-  "old:acronym_from_virtual:3255" => "CHMO",
-  "old:acronym_from_virtual:3204" => "PORO",
-  "old:acronym_from_virtual:1043" => "REX",
-  "old:acronym_from_virtual:1689" => "FYPO",
-  "old:acronym_from_virtual:3008" => "CARELEX",
-  "old:acronym_from_virtual:1570" => "TM-CONST",
-  "old:acronym_from_virtual:2070" => "obi-fged",
-  "old:acronym_from_virtual:2002" => "BRO-AreaOfResearch",
-  "old:acronym_from_virtual:2020" => "SNOMEDCT-ANAT",
-  "old:acronym_from_virtual:2092" => "envo_153",
-  "old:acronym_from_virtual:3045" => "ICD9CM-KM",
-  "old:acronym_from_virtual:3108" => "CCONT",
-  "old:acronym_from_virtual:1438" => "MCBCC",
-  "old:acronym_from_virtual:1348" => "OMIM",
-  "old:acronym_from_virtual:1515" => "INO",
-  "old:acronym_from_virtual:3176" => "XEO",
-  "old:acronym_from_virtual:2009" => "goslim_plant",
-  "old:acronym_from_virtual:2078" => "H1_NMOABA",
-  "old:acronym_from_virtual:3183" => "IFAR",
-  "old:acronym_from_virtual:2046" => "ICF-d6",
-  "old:acronym_from_virtual:3186" => "I2B2-LOINC",
-  "old:acronym_from_virtual:3265" => "BNO",
-  "old:acronym_from_virtual:2110" => "DTVPrecision",
-  "old:acronym_from_virtual:1581" => "HLTHINDCTRS",
-  "old:acronym_from_virtual:3220" => "MSTDE-FRE",
-  "old:acronym_from_virtual:3009" => "MEO",
-  "old:acronym_from_virtual:1694" => "CPT-KM",
-  "old:acronym_from_virtual:1052" => "PRO-ONT",
-  "old:acronym_from_virtual:1290" => "ABA-AMB",
-  "old:acronym_from_virtual:3000" => "MIXSCV",
-  "old:acronym_from_virtual:1038" => "PSDS",
-  "old:acronym_from_virtual:1023" => "FBbi",
-  "old:acronym_from_virtual:1060" => "CTONT",
-  "old:acronym_from_virtual:1249" => "SBRO",
-  "old:acronym_from_virtual:1152" => "MAT",
-  "old:acronym_from_virtual:2101" => "test1",
-  "old:acronym_from_virtual:1083" => "NPO",
-  "old:acronym_from_virtual:1013" => "EVOC",
-  "old:acronym_from_virtual:1407" => "TEDDY",
-  "old:acronym_from_virtual:2113" => "bibliographic",
-  "old:acronym_from_virtual:2030" => "iceci-descriptor_for_intent",
-  "old:acronym_from_virtual:2084" => "OGMD_OWL",
-  "old:acronym_from_virtual:1053" => "FMA",
-  "old:acronym_from_virtual:1422" => "MEDDRA",
-  "old:acronym_from_virtual:2010" => "goslim_candida",
-  "old:acronym_from_virtual:3273" => "HRDO",
-  "old:acronym_from_virtual:3152" => "REXO",
-  "old:acronym_from_virtual:3222" => "RSA",
-  "old:acronym_from_virtual:3207" => "ATOL",
-  "old:acronym_from_virtual:1498" => "EDAM",
-  "old:acronym_from_virtual:1109" => "SO",
-  "old:acronym_from_virtual:1224" => "SYMP",
-  "old:acronym_from_virtual:3253" => "OntoVIP",
-  "old:acronym_from_virtual:1237" => "SITBAC",
-  "old:acronym_from_virtual:2042" => "ICF-d2",
-  "old:acronym_from_virtual:3263" => "OntoBioUSP",
-  "old:acronym_from_virtual:1649" => "OSHPD",
-  "old:acronym_from_virtual:3247" => "EDDA",
-  "old:acronym_from_virtual:1671" => "QIBO",
-  "old:acronym_from_virtual:1585" => "XCO",
-  "old:acronym_from_virtual:1006" => "CL",
-  "old:acronym_from_virtual:1424" => "NDDF",
-  "old:acronym_from_virtual:3114" => "RPO",
-  "old:acronym_from_virtual:1426" => "MDDB",
-  "old:acronym_from_virtual:3169" => "GLYCO",
-  "old:acronym_from_virtual:2065" => "obi-device",
-  "old:acronym_from_virtual:1192" => "CPTAC",
-  "old:acronym_from_virtual:1032" => "NCIT",
-  "old:acronym_from_virtual:2067" => "RCJI",
-  "old:acronym_from_virtual:1017" => "FB-CV",
-  "old:acronym_from_virtual:2014" => "goslim_yeast",
-  "old:acronym_from_virtual:3251" => "ADO",
-  "old:acronym_from_virtual:1081" => "TTO",
-  "old:acronym_from_virtual:3162" => "RETO",
-  "old:acronym_from_virtual:1362" => "HAO",
-  "old:acronym_from_virtual:3240" => "RNPRIO",
-  "old:acronym_from_virtual:3272" => "CARD",
-  "old:acronym_from_virtual:3217" => "CABRO",
-  "old:acronym_from_virtual:2069" => "H1_NMOABA_1_0",
-  "old:acronym_from_virtual:3208" => "IDQA",
-  "old:acronym_from_virtual:3228" => "ELIG",
-  "old:acronym_from_virtual:1651" => "SDO",
-  "old:acronym_from_virtual:3126" => "NIFSUBCELL",
-  "old:acronym_from_virtual:3157" => "GENETRIAL",
-  "old:acronym_from_virtual:1070" => "GO",
-  "old:acronym_from_virtual:2088" => "SNMD_BC",
-  "old:acronym_from_virtual:1414" => "OGMS",
-  "old:acronym_from_virtual:1516" => "ICD10",
-  "old:acronym_from_virtual:1429" => "ICPC2P",
-  "old:acronym_from_virtual:3241" => "UDEF",
-  "old:acronym_from_virtual:2034" => "iceci-intent",
-  "old:acronym_from_virtual:2074" => "H1_NMOABA_2_0",
-  "old:acronym_from_virtual:1024" => "LHN",
-  "old:acronym_from_virtual:3003" => "CNO",
-  "old:acronym_from_virtual:2083" => "H1_NMOABA_4_1",
-  "old:acronym_from_virtual:1039" => "PROPREO",
-  "old:acronym_from_virtual:1638" => "ONTODM-CORE",
-  "old:acronym_from_virtual:2008" => "goslim_goa",
-  "old:acronym_from_virtual:3250" => "ONL-MR-DA",
-  "old:acronym_from_virtual:3136" => "NHDS",
-  "old:acronym_from_virtual:1393" => "IAO",
-  "old:acronym_from_virtual:1332" => "BFO",
-  "old:acronym_from_virtual:3203" => "HINO",
-  "old:acronym_from_virtual:2122" => "SciRes",
-  "old:acronym_from_virtual:2013" => "goslim_pombe",
-  "old:acronym_from_virtual:1627" => "HOMERUN",
-  "old:acronym_from_virtual:2117" => "event",
-  "old:acronym_from_virtual:1190" => "OPL",
-  "old:acronym_from_virtual:3002" => "MF",
-  "old:acronym_from_virtual:1417" => "FLU",
-  "old:acronym_from_virtual:1314" => "CLO",
-  "old:acronym_from_virtual:1101" => "ICD9CM",
-  "old:acronym_from_virtual:3029" => "SPTO",
-  "old:acronym_from_virtual:1381" => "NIFDYS",
-  "old:acronym_from_virtual:1440" => "GFO-BIO",
-  "old:acronym_from_virtual:1344" => "ICPC",
-  "old:acronym_from_virtual:1351" => "MESH",
-  "old:acronym_from_virtual:1094" => "PTRANS",
-  "old:acronym_from_virtual:1115" => "YPO",
-  "old:acronym_from_virtual:2126" => "CPT-mod",
-  "old:acronym_from_virtual:1621" => "NBO",
-  "old:acronym_from_virtual:3007" => "UCSFICU",
-  "old:acronym_from_virtual:2041" => "ICF-d1",
-  "old:acronym_from_virtual:2080" => "H1_NMOABA_4_0",
-  "old:acronym_from_virtual:1501" => "NEOMARK3",
-  "old:acronym_from_virtual:3158" => "SWEET",
-  "old:acronym_from_virtual:1335" => "PEO",
-  "old:acronym_from_virtual:1571" => "TM-OTHER-FACTORS",
-  "old:acronym_from_virtual:1061" => "SOPHARM",
-  "old:acronym_from_virtual:1411" => "ICF",
-  "old:acronym_from_virtual:1394" => "SSO",
-  "old:acronym_from_virtual:2018" => "SNOMEDCT-CF",
-  "old:acronym_from_virtual:1676" => "RCTONT",
-  "old:acronym_from_virtual:1086" => "OGDI",
-  "old:acronym_from_virtual:1699" => "VIVO",
-  "old:acronym_from_virtual:2079" => "ECGT",
-  "old:acronym_from_virtual:2082" => "H3_HIPSeptotemporalAxis",
-  "old:acronym_from_virtual:1321" => "NEMO",
-  "old:acronym_from_virtual:3267" => "DCO",
-  "old:acronym_from_virtual:2001" => "BRO-Activity",
-  "old:acronym_from_virtual:2072" => "NCBI_NMOsp_1_4_0",
-  "old:acronym_from_virtual:1343" => "HL7",
-  "old:acronym_from_virtual:1630" => "UCSFEPIC",
-  "old:acronym_from_virtual:2119" => "foaf",
-  "old:acronym_from_virtual:3094" => "DWC-TEST",
-  "old:acronym_from_virtual:3201" => "BCO",
-  "old:acronym_from_virtual:1105" => "MS",
-  "old:acronym_from_virtual:3199" => "UCSFICD910CM",
-  "old:acronym_from_virtual:1085" => "OGMD",
-  "old:acronym_from_virtual:1494" => "TMA",
-  "old:acronym_from_virtual:1484" => "ICECI",
-  "old:acronym_from_virtual:3259" => "InterNano",
-  "old:acronym_from_virtual:1444" => "CHEMINF",
-  "old:acronym_from_virtual:2048" => "ICF-d8",
-  "old:acronym_from_virtual:2051" => "ICF-d840-d859",
-  "old:acronym_from_virtual:3226" => "SBOL",
-  "old:acronym_from_virtual:2063" => "UCare_Demographics",
-  "old:acronym_from_virtual:1134" => "BT",
-  "old:acronym_from_virtual:1430" => "AI-RHEUM",
-  "old:acronym_from_virtual:1528" => "HUGO",
-  "old:acronym_from_virtual:3025" => "CTX",
-  "old:acronym_from_virtual:1131" => "MO",
-  "old:acronym_from_virtual:1404" => "UBERON",
-  "old:acronym_from_virtual:3194" => "PHENOSCAPE-EXT",
-  "old:acronym_from_virtual:1526" => "CRISP",
-  "old:acronym_from_virtual:2116" => "dcterms",
-  "old:acronym_from_virtual:3233" => "CHD",
-  "old:acronym_from_virtual:2058" => "SNOMEDCT-MAS",
-  "old:acronym_from_virtual:3270" => "suicideo",
-  "old:acronym_from_virtual:1037" => "PTO",
-  "old:acronym_from_virtual:3269" => "BICSO",
-  "old:acronym_from_virtual:1110" => "TAO",
-  "old:acronym_from_virtual:1011" => "IEV",
-  "old:acronym_from_virtual:1341" => "COSTART",
-  "old:acronym_from_virtual:2019" => "SNOMEDCT-TF",
-  "old:acronym_from_virtual:3043" => "MEDABBS",
-  "old:acronym_from_virtual:2091" => "NDF-RT",
-  "old:acronym_from_virtual:1583" => "CMO",
-  "old:acronym_from_virtual:1539" => "NIGO",
-  "old:acronym_from_virtual:3016" => "ERO",
-  "old:acronym_from_virtual:2031" => "iceci-details_for_mechanism",
-  "old:acronym_from_virtual:1507" => "CCO",
-  "old:acronym_from_virtual:1076" => "OCRE",
-  "old:acronym_from_virtual:1078" => "BSPO",
-  "old:acronym_from_virtual:3218" => "TRON",
-  "old:acronym_from_virtual:1135" => "pseudo",
-  "old:acronym_from_virtual:1544" => "CO",
-  "old:acronym_from_virtual:3246" => "ONL-MSA",
-  "old:acronym_from_virtual:2017" => "BRO-Chinese",
-  "old:acronym_from_virtual:1439" => "GFO",
-  "old:acronym_from_virtual:1015" => "FB-BT",
-  "old:acronym_from_virtual:1041" => "PSIMOD",
-  "old:acronym_from_virtual:3232" => "SNMI",
-  "old:acronym_from_virtual:1064" => "FB-SP",
-  "old:acronym_from_virtual:1655" => "PHARMGKB",
-  "old:acronym_from_virtual:2111" => "documentStatus",
-  "old:acronym_from_virtual:1014" => "FIX",
-  "old:acronym_from_virtual:1521" => "NEUMORE",
-  "old:acronym_from_virtual:3181" => "ONTOPNEUMO",
-  "old:acronym_from_virtual:2040" => "SNOMEDCT-CORE",
-  "old:acronym_from_virtual:1632" => "WSIO",
-  "old:acronym_from_virtual:2127" => "NIAID-GSC-BRC",
-  "old:acronym_from_virtual:1639" => "EPILONT",
-  "old:acronym_from_virtual:3205" => "ICD0",
-  "old:acronym_from_virtual:3236" => "VTO",
-  "old:acronym_from_virtual:2047" => "ICF-d7",
-  "old:acronym_from_virtual:1489" => "OAE",
-  "old:acronym_from_virtual:2115" => "dcelements",
-  "old:acronym_from_virtual:1588" => "ONTODT",
-  "old:acronym_from_virtual:3242" => "NHSQI",
-  "old:acronym_from_virtual:3197" => "HIMC-CPT",
-  "old:acronym_from_virtual:1027" => "MFO",
-  "old:acronym_from_virtual:1311" => "IDOMAL",
-  "old:acronym_from_virtual:3224" => "UCSFI9I10CMPCS",
-  "old:acronym_from_virtual:1149" => "DERMLEX",
-  "old:acronym_from_virtual:1012" => "ECO",
-  "old:acronym_from_virtual:2026" => "NCIt-Activity",
-  "old:acronym_from_virtual:1035" => "PW",
-  "old:acronym_from_virtual:1108" => "PAE",
-  "old:acronym_from_virtual:2138" => "Radlex3.9.1",
-  "old:acronym_from_virtual:2136" => "DILIo",
-  "old:acronym_from_virtual:1572" => "TM-SIGNS-AND-SYMPTS",
-  "old:acronym_from_virtual:3230" => "EPSO",
-  "old:acronym_from_virtual:1510" => "CPTH",
-  "old:acronym_from_virtual:2052" => "ICF-d860-d879",
-  "old:acronym_from_virtual:1107" => "PATO",
-  "old:acronym_from_virtual:1005" => "BTO",
-  "old:acronym_from_virtual:1580" => "AERO",
-  "old:acronym_from_virtual:3159" => "VARIO",
-  "old:acronym_from_virtual:2062" => "HOM-UCSF_UCareDispostion",
-  "old:acronym_from_virtual:1574" => "VHOG",
-  "old:acronym_from_virtual:1031" => "MPATH",
-  "old:acronym_from_virtual:1529" => "HCPCS",
-  "old:acronym_from_virtual:3058" => "DWC",
-  "old:acronym_from_virtual:1569" => "HPIO",
-  "old:acronym_from_virtual:2100" => "HSDB_OCRe",
-  "old:acronym_from_virtual:1491" => "IMGT-ONTOLOGY",
-  "old:acronym_from_virtual:2125" => "HCPCS-mod",
-  "old:acronym_from_virtual:3120" => "PCO",
-  "old:acronym_from_virtual:1019" => "FAO",
-  "old:acronym_from_virtual:2043" => "ICF-d3",
-  "old:acronym_from_virtual:2049" => "ICF-d9",
-  "old:acronym_from_virtual:3139" => "MIRNAO",
-  "old:acronym_from_virtual:1352" => "NDFRT",
-  "old:acronym_from_virtual:1044" => "SEP",
-  "old:acronym_from_virtual:1130" => "ACGT-MO",
-  "old:acronym_from_virtual:3214" => "OGSF",
-  "old:acronym_from_virtual:3179" => "GENE-CDS",
-  "old:acronym_from_virtual:1425" => "ICD10PCS",
-  "old:acronym_from_virtual:1090" => "AAO",
-  "old:acronym_from_virtual:1030" => "TGMA",
-  "old:acronym_from_virtual:1007" => "CHEBI",
-  "old:acronym_from_virtual:1087" => "OGR",
-  "old:acronym_from_virtual:3019" => "RH-MESH",
-  "old:acronym_from_virtual:2037" => "iceci-substance_use",
-  "old:acronym_from_virtual:3137" => "ONSTR",
-  "old:acronym_from_virtual:3216" => "BAO-GPCR",
-  "old:acronym_from_virtual:2064" => "LOINC",
-  "old:acronym_from_virtual:3198" => "SEMPHYSKB-HUMAN",
-  "old:acronym_from_virtual:1586" => "OntoOrpha",
-  "old:acronym_from_virtual:3192" => "MCCV",
-  "old:acronym_from_virtual:2104" => "MDRFRE",
-  "old:acronym_from_virtual:3078" => "PHENX",
-  "old:acronym_from_virtual:3020" => "CPO",
-  "old:acronym_from_virtual:1148" => "BP-METADATA",
-  "old:acronym_from_virtual:1021" => "EHDAA",
-  "old:acronym_from_virtual:3221" => "DERMO",
-  "old:acronym_from_virtual:3012" => "NONRCTO",
-  "old:acronym_from_virtual:1222" => "APO",
-  "old:acronym_from_virtual:1126" => "FHHO",
-  "old:acronym_from_virtual:1552" => "REPO",
-  "old:acronym_from_virtual:2066" => "RXNORM",
-  "old:acronym_from_virtual:3200" => "ZIP5",
-  "old:acronym_from_virtual:3249" => "ONL-DP",
-  "old:acronym_from_virtual:3190" => "PATHLEX",
-  "old:acronym_from_virtual:1672" => "DIKB",
-  "old:acronym_from_virtual:3210" => "TRAK",
-  "old:acronym_from_virtual:3266" => "OBI_BCGO",
-  "old:acronym_from_virtual:3244" => "SEDI",
-  "old:acronym_from_virtual:1369" => "PHYFIELD",
-  "old:acronym_from_virtual:3189" => "TOP-MENELAS",
-  "old:acronym_from_virtual:2106" => "ICF-NoCodeLabel",
-  "old:acronym_from_virtual:2112" => "geopolitical",
-  "old:acronym_from_virtual:2121" => "skos",
-  "old:acronym_from_virtual:1257" => "MEGO",
-  "old:acronym_from_virtual:1397" => "GAZ",
-  "old:acronym_from_virtual:3062" => "I2B2-PATVISDIM",
-  "old:acronym_from_virtual:1568" => "AEO",
-  "old:acronym_from_virtual:2022" => "BRO-Resource",
-  "old:acronym_from_virtual:1650" => "UNITSONT",
-  "old:acronym_from_virtual:3119" => "OBIWS",
-  "old:acronym_from_virtual:1578" => "ELIXHAUSER",
-  "old:acronym_from_virtual:1082" => "GRO",
-  "old:acronym_from_virtual:1490" => "PLATSTG",
-  "old:acronym_from_virtual:2006" => "FMA-SUBSET",
-  "old:acronym_from_virtual:1122" => "SPO",
-  "old:acronym_from_virtual:1504" => "CPT",
-  "old:acronym_from_virtual:1128" => "CDAO",
-  "old:acronym_from_virtual:1533" => "BAO",
-  "old:acronym_from_virtual:3237" => "VBCV",
-  "old:acronym_from_virtual:3209" => "OPE",
-  "old:acronym_from_virtual:1000" => "MA",
-  "old:acronym_from_virtual:2015" => "SOFA",
-  "old:acronym_from_virtual:2000" => "BRO-Core",
-  "old:acronym_from_virtual:1416" => "SCTSPA",
-  "old:acronym_from_virtual:2073" => "ECGF",
-  "old:acronym_from_virtual:2021" => "SNOMEDCT-ORG",
-  "old:acronym_from_virtual:1506" => "GO-EXT",
-  "old:acronym_from_virtual:1401" => "ICNP",
-  "old:acronym_from_virtual:2059" => "HOM-ICD9",
-  "old:acronym_from_virtual:1696" => "SYN",
-  "old:acronym_from_virtual:1058" => "SNPO",
-  "old:acronym_from_virtual:3195" => "ICD09",
-  "old:acronym_from_virtual:2027" => "3.0_OWL",
-  "old:acronym_from_virtual:1582" => "CAO",
-  "old:acronym_from_virtual:1048" => "WB-BT",
-  "old:acronym_from_virtual:1415" => "CTCAE",
-  "old:acronym_from_virtual:1523" => "OBOE-SBC",
-  "old:acronym_from_virtual:3185" => "GPI",
-  "old:acronym_from_virtual:2004" => "NeuroFMA",
-  "old:acronym_from_virtual:1022" => "EHDA",
-  "old:acronym_from_virtual:1419" => "TAXRANK",
-  "old:acronym_from_virtual:2024" => "RadLex_OWL",
-  "old:acronym_from_virtual:2139" => "PROV",
-  "old:acronym_from_virtual:3274" => "MSV",
-  "old:acronym_from_virtual:1509" => "ICPS",
-  "old:acronym_from_virtual:1626" => "EMO",
-  "old:acronym_from_virtual:1036" => "PECO",
-  "old:acronym_from_virtual:1614" => "IXNO",
-  "old:acronym_from_virtual:3030" => "CANCO",
-  "old:acronym_from_virtual:2118" => "FaBiO",
-  "old:acronym_from_virtual:1487" => "ICD11-BODYSYSTEM",
-  "old:acronym_from_virtual:1665" => "SHR",
-  "old:acronym_from_virtual:1084" => "NIFSTD",
-  "old:acronym_from_virtual:2057" => "SNOMED-Ethnic-Group",
-  "old:acronym_from_virtual:1141" => "OPB",
-  "old:acronym_from_virtual:3227" => "OVAE",
-  "old:acronym_from_virtual:1029" => "IMR",
-  "old:acronym_from_virtual:1353" => "SNOMEDCT",
-  "old:acronym_from_virtual:2102" => "WHOFRE",
-  "old:acronym_from_virtual:1132" => "NCBITAXON",
-  "old:acronym_from_virtual:2028" => "iceci-activity",
-  "old:acronym_from_virtual:1538" => "ROLEO",
-  "old:acronym_from_virtual:2044" => "ICF-d4",
-  "old:acronym_from_virtual:1088" => "MHC",
-  "old:acronym_from_virtual:3167" => "GLOB",
-  "old:acronym_from_virtual:2081" => "H2_HIPClassicRegions",
-  "old:acronym_from_virtual:3147" => "CLIN-EVAL",
-  "old:acronym_from_virtual:3006" => "OOEVV",
-  "old:acronym_from_virtual:3180" => "MEDO",
-  "old:acronym_from_virtual:1505" => "OMIT",
-  "old:acronym_from_virtual:1553" => "ICD10CM",
-  "old:acronym_from_virtual:2107" => "BIOA3v",
-  "old:acronym_from_virtual:3046" => "MDCDRG",
-  "old:acronym_from_virtual:1633" => "COGAT",
-  "old:acronym_from_virtual:1659" => "VT",
-  "old:acronym_from_virtual:1040" => "PPIO",
-  "old:acronym_from_virtual:3129" => "CONSENT-ONT",
-  "old:acronym_from_virtual:1008" => "DDANAT",
-  "old:acronym_from_virtual:1020" => "HC",
-  "old:acronym_from_virtual:1150" => "RS"
-}
