--- lib/resolver/virtual_from_acronym.rb
@@ -1,533 +1,0 @@
-BpidResolver::VIRTUAL_FROM_ACRONYM = {
-  "JERM" => 1488,
-  "OGSF" => 3214,
-  "NATPRO" => 3004,
-  "citation" => 2114,
-  "ATC" => 3021,
-  "SNMD_BC" => 2088,
-  "WHO-ART" => 1354,
-  "ICD9CM-PROC" => 2124,
-  "OGR" => 1087,
-  "BAO" => 1533,
-  "ATMO" => 1099,
-  "goslim_candida" => 2010,
-  "CHEBI" => 1007,
-  "SNPO" => 1058,
-  "provenance" => 2120,
-  "WHOFRE" => 2102,
-  "EMAP" => 1010,
-  "CPT-mod" => 2126,
-  "CTONT" => 1060,
-  "BICSO" => 3269,
-  "NMR" => 1033,
-  "NPO" => 1083,
-  "NEOMARK4" => 1686,
-  "BRO" => 1104,
-  "TMA" => 1494,
-  "GLYCANONT" => 3231,
-  "ICF-d5" => 2045,
-  "GLOB" => 3167,
-  "SNOMED-Ethnic-Group" => 2057,
-  "SBRO" => 1249,
-  "YPO" => 1115,
-  "PO" => 1587,
-  "CPT" => 1504,
-  "EXO" => 1575,
-  "geopolitical" => 2112,
-  "NIC" => 1428,
-  "CCONT" => 3108,
-  "CLIN-EVAL" => 3147,
-  "MMO" => 1584,
-  "AMINO-ACID" => 1054,
-  "OBOREL" => 1042,
-  "IDOBRU" => 1537,
-  "ICF-d860-d879" => 2052,
-  "BP-METADATA" => 1148,
-  "QUDT" => 3031,
-  "HOM-UCSF_UCareDispostion" => 2062,
-  "MSTDE-FRE" => 3220,
-  "SNOMEDCT-ORG" => 2021,
-  "NCBI_NMOsp_1_4_0" => 2072,
-  "HP" => 1125,
-  "BTO" => 1005,
-  "UO" => 1112,
-  "SWEET" => 3158,
-  "OGMD" => 1085,
-  "VSAO" => 1555,
-  "ICF-d6" => 2046,
-  "RPO" => 3114,
-  "ECO" => 1012,
-  "HCPCS" => 1529,
-  "MIRO" => 1077,
-  "BSAO" => 3271,
-  "BiositemapIM" => 2003,
-  "ooevv-vaccine" => 2109,
-  "COGPO" => 1560,
-  "CMS" => 3146,
-  "CCO" => 1507,
-  "NHDS" => 3136,
-  "pseudo" => 1135,
-  "TGMA" => 1030,
-  "RID" => 1260,
-  "TM-SIGNS-AND-SYMPTS" => 1572,
-  "HAO" => 1362,
-  "VIVO" => 1699,
-  "EPSO" => 3230,
-  "OntoBioUSP" => 3263,
-  "suicideo" => 3270,
-  "SDO" => 1651,
-  "ABA-AMB" => 1290,
-  "HPIO" => 1569,
-  "AAO" => 1090,
-  "MCCL" => 1541,
-  "HOM-TEST" => 3038,
-  "HIMC-CPT" => 3197,
-  "ACGT-MO" => 1130,
-  "REX" => 1043,
-  "TM-MER" => 1573,
-  "PHENOSCAPE-EXT" => 3194,
-  "IEV" => 1011,
-  "EHDA" => 1022,
-  "FBbi" => 1023,
-  "CHEMBIO" => 1615,
-  "ICD10" => 1516,
-  "ROLEO" => 1538,
-  "GFO" => 1439,
-  "MF" => 3002,
-  "IMMDIS" => 3127,
-  "dcelements" => 2115,
-  "ADW" => 1530,
-  "BOF" => 3234,
-  "iceci-countermeasures" => 2029,
-  "GALEN" => 1055,
-  "PAE" => 1108,
-  "OBI" => 1123,
-  "HOM" => 1328,
-  "HOMERUN" => 1627,
-  "CABRO" => 3217,
-  "ICD9CM" => 1101,
-  "SNOMEDCT-MAS" => 2058,
-  "ICF" => 1411,
-  "WH" => 3264,
-  "CHMO" => 3255,
-  "H1_NMOABA_1_0" => 2069,
-  "BSPO" => 1078,
-  "MDRFRE" => 2104,
-  "ICF-NoCodeLabel" => 2106,
-  "GRO" => 1082,
-  "NCIT" => 1032,
-  "UCSFICD910CM" => 3199,
-  "ADO" => 3251,
-  "BDO" => 1613,
-  "NeuroFMA" => 2004,
-  "MCBCC" => 1438,
-  "ICD09" => 3195,
-  "ICF-d840-d859%2Fd860-d879" => 2053,
-  "RCTONT" => 1676,
-  "ERNO" => 3268,
-  "ONTODM-KDD" => 3077,
-  "IXNO" => 1614,
-  "ICNP" => 1401,
-  "KISAO" => 1410,
-  "EDDA" => 3247,
-  "GFO-BIO" => 1440,
-  "FaBiO" => 2118,
-  "UCare_Demographics" => 2063,
-  "OGMS" => 1414,
-  "Radlex3.9.1" => 2138,
-  "MEO" => 3009,
-  "SPTO" => 3029,
-  "CHEMINF" => 1444,
-  "CANONT" => 3178,
-  "HC" => 1020,
-  "PPIO" => 1040,
-  "H1_NMOABA" => 2078,
-  "EHDAA" => 1021,
-  "VSO" => 3124,
-  "NIGO" => 1539,
-  "SPO" => 1122,
-  "VTO" => 3236,
-  "MP" => 1025,
-  "GCC" => 3017,
-  "BILA" => 1114,
-  "iceci-descriptor_for_intent" => 2030,
-  "obi-device" => 2065,
-  "CL" => 1006,
-  "RH-MESH" => 3019,
-  "ICD10CM" => 1553,
-  "BRIDG" => 3150,
-  "UCSFI9I10CMPCS" => 3224,
-  "MIXS" => 3215,
-  "SNOMEDCT-CF" => 2018,
-  "FMA-SUBSET" => 2006,
-  "PTO" => 1037,
-  "CMO" => 1583,
-  "FB-SP" => 1064,
-  "EP" => 1142,
-  "SIO" => 1532,
-  "InterNano" => 3259,
-  "ICF-d1" => 2041,
-  "REPO" => 1552,
-  "PHARE" => 1550,
-  "UNITSONT" => 1650,
-  "REXO" => 3152,
-  "DOID" => 1009,
-  "ICDO3" => 1670,
-  "DIAGONT" => 3013,
-  "I2B2-PATVISDIM" => 3062,
-  "NCBI_NMOsp_1_2_1" => 2055,
-  "MDDB" => 1426,
-  "PROV" => 2139,
-  "ICF-d810-d839" => 2050,
-  "FLU" => 1417,
-  "DC-CL" => 1144,
-  "LIPRO" => 1183,
-  "UCSFICU" => 3007,
-  "SNOMEDCT-CORE" => 2040,
-  "NCIt-Activity" => 2026,
-  "CARO" => 1063,
-  "CPT-KM" => 1694,
-  "MFO" => 1027,
-  "GRO-CPGA" => 1001,
-  "RETO" => 3162,
-  "PATHLEX" => 3190,
-  "GEOSPECIES" => 1247,
-  "OAE" => 1489,
-  "EMO" => 1626,
-  "MPO" => 3191,
-  "OMIM" => 1348,
-  "CTX" => 3025,
-  "FAO" => 1019,
-  "goslim_plant" => 2009,
-  "NIFSTD" => 1084,
-  "iceci-substance_use" => 2037,
-  "BIOA3v" => 2107,
-  "QIBO" => 1671,
-  "iceci-details_for_activity" => 2038,
-  "BCTEO" => 3262,
-  "CANCO" => 3030,
-  "PDQ" => 1349,
-  "NDDF" => 1424,
-  "VO" => 1172,
-  "H1_NMOABA_4_1" => 2083,
-  "iceci-intent" => 2034,
-  "IAO" => 1393,
-  "RS" => 1150,
-  "ONL-DP" => 3249,
-  "SciRes" => 2122,
-  "CARD" => 3272,
-  "goslim_generic_1_770" => 2007,
-  "EVOC" => 1013,
-  "MSHSPA_1" => 1420,
-  "SNOMEDCT-TF" => 2019,
-  "MAT" => 1152,
-  "GRO-CPD" => 1047,
-  "GLYCO" => 3169,
-  "UCSFORTHO" => 1697,
-  "PCO" => 3120,
-  "CRISP" => 1526,
-  "IDODEN" => 3174,
-  "iceci-details_for_place_of_occurrence" => 2032,
-  "ICD0" => 3205,
-  "PVONTO" => 1567,
-  "GEXO" => 3151,
-  "DIKB" => 1672,
-  "FB-CV" => 1017,
-  "SBO" => 1046,
-  "PEO" => 1335,
-  "Amino_Acid-zh_CN" => 1168,
-  "RNAO" => 1500,
-  "PHYLONT" => 1616,
-  "GO-EXT" => 1506,
-  "ICF-d9" => 2049,
-  "ONL-MSA" => 3246,
-  "NIFSUBCELL" => 3126,
-  "PR" => 1062,
-  "CAO" => 1582,
-  "VBCV" => 3237,
-  "ICD9CM-KM" => 3045,
-  "FIX" => 1014,
-  "ICECI" => 1484,
-  "NEUMORE" => 1521,
-  "WSIO" => 1632,
-  "goslim_goa" => 2008,
-  "ICF-d3" => 2043,
-  "MDCDRG" => 3046,
-  "DWC-TEST" => 3094,
-  "BT" => 1134,
-  "H3_HIPSeptotemporalAxis" => 2082,
-  "MSV" => 3274,
-  "DILIo" => 2136,
-  "ICF-d840-d859" => 2051,
-  "MCCV" => 3192,
-  "HL7" => 1343,
-  "XCO" => 1585,
-  "ICF-d7" => 2047,
-  "ZIP5" => 3200,
-  "SSO" => 1394,
-  "RXNORM" => 2066,
-  "DERMO" => 3221,
-  "obi-fged" => 2070,
-  "RadLex_OWL" => 2024,
-  "H1_NMOABA_2_0" => 2074,
-  "MEDDRA" => 1422,
-  "BRO-Resource" => 2022,
-  "TMO" => 1461,
-  "LHN" => 1024,
-  "PROPREO" => 1039,
-  "CO-WHEAT" => 1545,
-  "CARELEX" => 3008,
-  "NDFRT" => 1352,
-  "iceci-instrument_object_substance" => 2033,
-  "PRO-ONT" => 1052,
-  "TADS" => 1065,
-  "HLTHINDCTRS" => 1581,
-  "WB-PHENOTYPE" => 1067,
-  "OGDI" => 1086,
-  "TEO" => 3042,
-  "PHYFIELD" => 1369,
-  "NEMO" => 1321,
-  "CBO" => 1158,
-  "CPO" => 3020,
-  "SNOMEDCT" => 1353,
-  "DWC" => 3058,
-  "foaf" => 2119,
-  "LDA" => 1398,
-  "HUGO" => 1528,
-  "AI-RHEUM" => 1430,
-  "MHC" => 1088,
-  "XAO" => 1095,
-  "UDEF" => 3241,
-  "ICF-d8" => 2048,
-  "H1_NMOABA_4_0" => 2080,
-  "test1" => 2101,
-  "TM-CONST" => 1570,
-  "FB-DV" => 1016,
-  "PORO" => 3204,
-  "TTO" => 1081,
-  "ECGF" => 2073,
-  "EHDAA2" => 1517,
-  "BCGO" => 1304,
-  "DERMLEX" => 1149,
-  "CO" => 1544,
-  "ICPCFRE" => 2103,
-  "TOP-MENELAS" => 3189,
-  "SMD_NPLSM_BRST" => 2086,
-  "OntoOrpha" => 1586,
-  "GAZ" => 1397,
-  "ELIXHAUSER" => 1578,
-  "CPTAC" => 1192,
-  "ONTOKBCF" => 3155,
-  "TAO" => 1110,
-  "UCSFXPLANT" => 1661,
-  "OBI_BCGO" => 3266,
-  "MEGO" => 1257,
-  "VARIO" => 3159,
-  "VT" => 1659,
-  "SYMP" => 1224,
-  "BRO-Activity" => 2001,
-  "SOY" => 3028,
-  "FHHO" => 1126,
-  "OPL" => 1190,
-  "iceci-mechanism" => 2035,
-  "IFAR" => 3183,
-  "NHSQI" => 3242,
-  "LOINC" => 2064,
-  "IDO" => 1092,
-  "OBIWS" => 3119,
-  "EFO" => 1136,
-  "ONTODT" => 1588,
-  "ONTOMA" => 3090,
-  "ICD10PCS" => 1425,
-  "SEDI" => 3244,
-  "RSA" => 3222,
-  "GENETRIAL" => 3157,
-  "SCTSPA" => 1416,
-  "iceci-details_for_mechanism" => 2031,
-  "SAO" => 1068,
-  "NIFDYS" => 1381,
-  "ENVO" => 1069,
-  "INO" => 1515,
-  "MA" => 1000,
-  "MFOEM" => 1666,
-  "BRO-Chinese" => 2017,
-  "NCBITAXON" => 1132,
-  "HCPCS-mod" => 2125,
-  "dcterms" => 2116,
-  "I2B2-LOINC" => 3186,
-  "EPICMEDS" => 3032,
-  "BRO-Core" => 2000,
-  "CDAO" => 1128,
-  "ICF-d4" => 2044,
-  "TEST-PROD" => 3211,
-  "BAO-GPCR" => 3216,
-  "ooevv-tractTrace" => 2108,
-  "PATO" => 1107,
-  "OBOE-SBC" => 1523,
-  "BP" => 1522,
-  "OCRE" => 1076,
-  "skos" => 2121,
-  "MIXSCV" => 3000,
-  "SPD" => 1091,
-  "BHO" => 1116,
-  "STNFRDRXDEMO" => 3261,
-  "FB-BT" => 1015,
-  "NCBI_NMOsp_1_2_2" => 2056,
-  "WB-LS" => 1049,
-  "SNOMEDCT-ANAT" => 2020,
-  "TRAK" => 3210,
-  "IDOMAL" => 1311,
-  "SMD_NPLSM" => 2085,
-  "goslim_pombe" => 2013,
-  "SSE" => 3252,
-  "CONSENT-ONT" => 3129,
-  "BCO" => 3201,
-  "RCD" => 1427,
-  "AEO" => 1568,
-  "TAXRANK" => 1419,
-  "PHENX" => 3078,
-  "goslim_pir" => 2012,
-  "MO" => 1131,
-  "PSDS" => 1038,
-  "ONTODM-CORE" => 1638,
-  "CHD" => 3233,
-  "PHARMGKB" => 1655,
-  "NIF-RTH" => 2025,
-  "OPB" => 1141,
-  "MSHFRE" => 1456,
-  "ZEA" => 1050,
-  "MEDLINEPLUS" => 1347,
-  "SO" => 1109,
-  "OBI_IEDB_view" => 2093,
-  "PSIMOD" => 1041,
-  "GPI" => 3185,
-  "IMGT-ONTOLOGY" => 1491,
-  "ONL-MR-DA" => 3250,
-  "SHR" => 1665,
-  "NHSQI2009" => 3243,
-  "FDA-MEDDEVICE" => 1576,
-  "OVAE" => 3227,
-  "NDF-RT" => 2091,
-  "CPRO" => 1059,
-  "CSSO" => 3239,
-  "event" => 2117,
-  "CLO" => 1314,
-  "ECG" => 1146,
-  "PMR" => 3015,
-  "IDQA" => 3208,
-  "OntoVIP" => 3253,
-  "ZFA" => 1051,
-  "HCPCS-HIMC" => 3206,
-  "OSHPD" => 1649,
-  "ICF-d2" => 2042,
-  "BIOMODELS" => 3022,
-  "PHENOMEBLAST" => 1656,
-  "OMIT" => 1505,
-  "SITBAC" => 1237,
-  "NIAID-GSC-BRC" => 2127,
-  "SNMI" => 3232,
-  "SOPHARM" => 1061,
-  "BNO" => 3265,
-  "DTVPrecision" => 2110,
-  "VANDF" => 1527,
-  "OOEVV" => 3006,
-  "HSDB_OCRe" => 2100,
-  "LSM" => 1520,
-  "BRO-AreaOfResearch" => 2002,
-  "WB-BT" => 1048,
-  "H2_HIPClassicRegions" => 2081,
-  "ERO" => 3016,
-  "EDAM" => 1498,
-  "HRDO" => 3273,
-  "OPE" => 3209,
-  "IMR" => 1029,
-  "SYN" => 1696,
-  "NONRCTO" => 3012,
-  "GCO" => 3223,
-  "RCJI" => 2067,
-  "GO" => 1070,
-  "OMRSE" => 1565,
-  "ONSTR" => 3137,
-  "SNMD_BDY" => 2105,
-  "BIRNLEX" => 1089,
-  "APO" => 1222,
-  "DDANAT" => 1008,
-  "ONTOPNEUMO" => 3181,
-  "NIFCELL" => 1402,
-  "SOFA" => 2015,
-  "ATO" => 1370,
-  "CLINIC" => 3092,
-  "PROVO" => 3131,
-  "NTDO" => 3153,
-  "RADLEX" => 1057,
-  "FYPO" => 1689,
-  "MAO" => 1026,
-  "MS" => 1105,
-  "VHOG" => 1574,
-  "PDO" => 3238,
-  "DDI" => 1540,
-  "PECO" => 1036,
-  "MESH" => 1351,
-  "COGAT" => 1633,
-  "CPTH" => 1510,
-  "MIRNAO" => 3139,
-  "goslim_yeast" => 2014,
-  "iceci-place_of_occurrence" => 2036,
-  "DCO" => 3267,
-  "FMA" => 1053,
-  "MPATH" => 1031,
-  "PTRANS" => 1094,
-  "ECGT" => 2079,
-  "bibliographic" => 2113,
-  "ATOL" => 3207,
-  "ELIG" => 3228,
-  "OGI" => 1100,
-  "USSOC" => 3104,
-  "GLYCOPROT" => 3212,
-  "AERO" => 1580,
-  "TEDDY" => 1407,
-  "HINO" => 3203,
-  "GENE-CDS" => 3179,
-  "TRON" => 3218,
-  "OBR-Scolio" => 3258,
-  "ZIP3" => 3184,
-  "OGMD_OWL" => 2084,
-  "ICPC2P" => 1429,
-  "SEP" => 1044,
-  "PW" => 1035,
-  "SBOL" => 3226,
-  "envo_153" => 2092,
-  "TOK" => 1418,
-  "COSTART" => 1341,
-  "ICPS" => 1509,
-  "UBERON" => 1404,
-  "NCBI_NMOsp_1_0_0" => 2054,
-  "PMA" => 1497,
-  "MEDABBS" => 3043,
-  "EPILONT" => 1639,
-  "NCBITaxonOWL" => 2023,
-  "BFO" => 1332,
-  "SEMPHYSKB-HUMAN" => 3198,
-  "ICD11-BODYSYSTEM" => 1487,
-  "MEDO" => 3180,
-  "CTCAE" => 1415,
-  "XEO" => 3176,
-  "NEOMARK3" => 1501,
-  "SWO" => 1413,
-  "SuicidO" => 3245,
-  "GWAS_EFO_SKOS" => 2099,
-  "documentStatus" => 2111,
-  "ICPC" => 1344,
-  "PEDTERM" => 1640,
-  "HOM-ICD9" => 2059,
-  "iceci-activity" => 2028,
-  "UCSFEPIC" => 1630,
-  "CNO" => 3003,
-  "PLATSTG" => 1490,
-  "3.0_OWL" => 2027,
-  "DEMOGRAPH" => 3047,
-  "TM-OTHER-FACTORS" => 1571,
-  "RNPRIO" => 3240,
-  "MSTDE" => 3219,
-  "NBO" => 1621
-}
